An webapplication that was created for a car show
Me and a group of classmates created a web application using the Laravel Framework. This was made for the final project in our Systems Development course. We had to choose a company and develop an application for them. So we decided to chose a car show company and develop a web application which allowed users to purchase tickets for the show (and they can get a cheaper ticket if they purchase it weeks in advance), register their vehicle to show off at the show, and create their own account. It also allows the employees of the car show to view the vehicles that got registered, view the tickets purchased, view normal users account information (except for their password).