Created a blackjack game using C++
Me and a group of classmates developed and created a blackjack game using the C++ programming language and Dev-C++. The user plays the blackjack game on the windows command line. This means that the game is fully textbased. We were able to create AI for the user to play against. But if the user had someone else that he or she wanted to play against. We were also able to have it be local multiplayer (the people playing against one another must use the same device). Finally we made it so that the dealer can pull a random card (not true randomness) from the deck.