Created a payroll system
For the final project in my .Net development course I created a payroll system for a company which pay's their employees per message they produce. So, I was able to create a tab in the application which is where the manager inputs the workers payment and if they are either a normal or senior employee. If the senior employee puts in more work, then the normal employee then they would be getting paid more. Then there is a tab called "Employee List" which shows the total amount of employees that were inputted and how much they were paid. The manager is also able to edit or delete any record on that list.