Final Project at the end of my college program
Me and a group of classmates took all our knowledge during our time in college and developed a virtual reality tech demo. Since we had time restraints (we needed to fully complete this project in three months). We could only make multiple different types of demos instead of a fully-fledged virtual reality video game. So, we decided to create a bow and arrow demo, a fruit ninja demo, a shooting range demo, and a basketball demo. We were able to get just about full functionality in all the demos but there was a couple of little issues at the end. The only thing that wasn't included was anything sound related (sound FX and music). We were able to create basic AI and I was able to create models for the bow and arrow demo as well as the fruit ninja demo. The guns in the shooting range demo was from a copyright free 3D modeler. Who had his/her work free to download and use.